🛠️refactor: PrivacySheet 컴포넌트 기능 hook으로 분리 - 버튼 배경색, radio 체크 박스 색상 gray 색상으로 변경
