+++ src/app/_components/client/PrivacySheet/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as usePrivacySheet } from "./usePrivacySheet/usePrivacySheet";
